+++ pvzclass/Enums/ChallengeScreen.hpp
@@ -1,0 +1,14 @@
+#pragma once
+namespace ChallengePage
+{
+	enum ChallengePage
+	{
+		SURVIVAL,
+		CHALLENGE,
+		LIMBO,
+		PUZZLE,
+	};
+
+	const char* ToString(ChallengePage page);
+
+}