+++ pvzclass/pvzclass.h
@@ -1,6 +1,7 @@
 ﻿#pragma once
 #include "ProcessOpener.h"
 #include "PVZ.h"
+#include "Classes.hpp"
 #include "Creators.h"
 #include "Draw.h"
 #include "Extensions.h"